remove old api retrieval endpoint
