+++ i18n/translations/fr/admin-templates.json
@@ -28,15 +28,9 @@
     "new": "Le formulaire a été créé avec succès",
     "formID": "Identifiant du formulaire :",
     "edit": "Modifier le formulaire :",
-    "bearerToken": {
-      "current": "Jeton d'accès :",
-      "refreshButton": "Renouveler le jeton d'accès",
-      "refreshError": "Le renouvellement du jeton d'accès n'a pas fonctionné.",
-      "currentAriaLabel": "Jeton d'accès"
-    },
+
     "tabLabels": {
       "jsonUpload": "Téléversement JSON",
-      "bearerToken": "Jeton d'accès",
       "formAccess": "Accès au formulaire"
     },
     "formAccess": {
