+++ lib/types/index.ts
@@ -43,13 +43,7 @@ export type {
   SearchParams,
 } from "./utility-types";
 
-export type {
-  BearerTokenPayload,
-  TemporaryTokenPayload,
-  BearerResponse,
-  VaultSubmission,
-  VaultSubmissionList,
-} from "./retrieval-types";
+export type { VaultSubmission, VaultSubmissionList } from "./retrieval-types";
 
 export { VaultStatus } from "./retrieval-types";
