+++ lib/types/retrieval-types.ts
@@ -1,19 +1,6 @@
 import { Responses } from "./form-response-types";
 import { TypeOmit } from ".";
 
-export interface BearerTokenPayload {
-  formID: string;
-}
-
-export interface TemporaryTokenPayload {
-  email: string;
-  formID: string;
-}
-
-export type BearerResponse = {
-  bearerToken: string;
-};
-
 export type VaultSubmission = {
   formID: string;
   submissionID: string;
