+++ package.json
@@ -76,7 +76,6 @@
     "immer": "10.0.3",
     "ioredis": "5.3.2",
     "jsonschema": "1.4.1",
-    "jsonwebtoken": "9.0.2",
     "jszip": "^3.10.1",
     "lexical": "^0.12.6",
     "lodash.debounce": "4.0.8",
@@ -131,7 +130,6 @@
     "@types/ioredis-mock": "^8.2.5",
     "@types/jest": "^29.5.11",
     "@types/json2md": "^1.5.4",
-    "@types/jsonwebtoken": "^8.5.9",
     "@types/lodash.debounce": "^4.0.9",
     "@types/lodash.isequal": "^4",
     "@types/lodash.set": "^4.3.9",
