+++ prisma/schema.prisma
@@ -56,6 +56,7 @@ model Template {
   isPublished       Boolean         @default(false)
   deliveryOption    DeliveryOption?
   securityAttribute String          @default("Protected A")
+  // @todo  Need to remove this field
   bearerToken       String?
   ttl               DateTime?
   apiUsers          ApiUser[]
