--- utils/retrievalAPI/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "retrieval_api",
-  "version": "0.1.0",
-  "description": "Retrieve form respones from the retrieval API",
-  "main": "retrieveResponse.ts",
-  "scripts": {
-    "retrieve": "tsx retrieveResponse.ts"
-  },
-  "keywords": [],
-  "author": "Bryan Robitaille",
-  "dependencies": {
-    "@types/node": "^18.19.3",
-    "csv-writer": "^1.6.0",
-    "tsx": "^4.7.0",
-    "typescript": "^5.3.3"
-  },
-  "packageManager": "yarn@4.0.2"
-}
